Refactor graph layering to merge CFG and MOP structures
* refactor graph layering into unified CFG abstraction
* address review nits in unified cfg traversal
