+++ rapx/src/graphs/mod.rs
@@ -1,4 +1,3 @@
 pub mod cfg;
-pub mod mop;
 pub mod scc;
 pub mod scc_paths;
